fix: curve fi interface

+++ src/interfaces/I3PoolCurve.sol
@@ -8,5 +8,5 @@ pragma solidity ^0.8.17;
 interface I3PoolCurve {
     function get_dy(int128 i, int128 j, uint256 dx) external view returns (uint256);
     function exchange(int128 i, int128 j, uint256 dx, uint256 min_dy) external;
-    function coins(int128 arg0) external view returns (address);
+    function coins(uint256 arg0) external view returns (address);
 }
